api/python/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.3.0
python-jose==3.3.0
